django_api/requirements/base.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2

# OpenAI Integration
openai==1.13.3
requests==2.31.0

# Async Task Queue (for OpenAI API calls)
celery==5.3.6
redis==5.0.1
django-celery-beat==2.6.0

# Validation
django-filter==23.5

# Utilities
python-dotenv==1.0.1
dj-database-url==2.1.0